v2: Fix InMemoryCache.Clear() concurrency
The sync.Map implementation of InMemoryCache contains a concurrency
bug around Clear() with concurrent calls to Get().

Protecting the value of `items` with a mutex only in Clear() is not
enough to prevent data races, as other methods do not respect that
mutex.

Adding a mutex around `items` in all the other Cache methods would
defeat the purpose of using sync.Map in the first place.

Instead of sync.Map, we use a regular map protected by mutex.
